+++ data-raw/La11.R
@@ -40,6 +40,58 @@ ecc <- readr::read_table("http://vo.imcce.fr/insola/earth/online/earth/La2010/La
 dplyr::glimpse(dat)
 dplyr::glimpse(ecc)
 
+# back-calculate the input columns
+# use wolframalpha to solve this system of equations:
+# just e and \varpi:
+# https://www.wolframalpha.com/input?i=k%3De%5Ccos%28%5Cvarpi%29%3Bh%3De%5Csin%28%5Cvarpi%29++solve+for+e%2C+%5Cvarpi
+# just i and \Omega
+# solve \varpi from e and k
+## https://www.wolframalpha.com/input?i=k%3De%5Ccos%28%5Cvarpi%29+solve+for+%5Cvarpi+
+# solve \varpi from e and h
+## https://www.wolframalpha.com/input?i=k%3De%5Ccos%28%5Cvarpi%29+solve+for+%5Cvarpi+
+# \varpi = -sin^-1(h/e)+2*pi*n+pi (n is in the set of integers)
+# vpi = asin(h/e)+2*pi*n+pi (n is in the set of integers)
+# \varpi = sin^-1(h/e)+2*pi*n (n is in the set of integers)
+# vpi = asin(h/e)+2*pi*n (n is in the set of integers)
+# https://www.wolframalpha.com/input?i=k%3De%5Ccos%28%5Cvarpi%29%3Bh%3De%5Csin%28%5Cvarpi%29%3Bq%3D%5Csin%28i%2F2%29%5Ccos%28%5COmega%29%3Bp%3D%5Csin%28i%2F2%29%5Csin%28%5COmega%29%2C+solve+for+i%2C+%5Cvarpi%2C+%5COmega
+
+n <- 1
+La11a <- left_join(dat, ecc) |>
+  mutate(
+    ## om = from ???
+    # wiki says \omega = \arccos(\frac{\vec{n} \dot e}{|n||e|})
+    # where n is a vector pointing towards the ascending node and e is eccentricity vector (vector pointing to periapsis)
+    om = NA_real_,
+    ## oom = from p and q,
+    oom = 2 * (pi * n + atan((sqrt(p^2 + q^2) - q) / p)) - 2 * pi,
+    ## i = from p and q,
+    i = 2 * (2 * pi * n - asin(q / cos(oom)) + pi),
+    ## vpi = from e + k and h,
+    ## vpi_h = asin(h/e) + 2 * pi * n, #*n (n is in the set of integers)
+    ## vpi_k = acos(k/e) + 2 * pi * n, #*n (n is in the set of integers)
+    vpi = 2 * pi * n -2 * atan((sqrt(h^2 + k^2) + k) / h) - 2 * pi,
+    ## mn = from mean longitude?
+    # wiki says: M = E - e sin(E) for which you need the Newton--Rhapson method
+    # mean longitude l = \Omega + \omega + M (mean longitude)
+    mn = l - oom - om # but we don't have om yet
+    # or use true anomaly \nu somehow?
+    # \cos\nu = \frac{\cos(E)-e}{1-e\cos(E)}
+    # \nu = E + 2*\arctan(\frac{\beta\sin(E)}{1-\beta\cos(E)})
+  ) |>
+  tidylog::select(t_ka, a, e, i, om, oom, vpi, l, mn, h, k, p, q) |>
+  mutate(solution = "La11c")
+
+write_rds(La11a, "~/SurfDrive/Postdoc1/prj/2023-03-23_snvec-R/out/La11.rds")
+# TODO: rotate everything so that it aligns with J2000 rather than invariant reference frame
+
+# the stuff below is older!
+
+
+
+
+
+
+
 ## rename some of the names in dat
 ## to make the naming consistent with the C code
 dat <- dat |>